requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
